+++ src/main/java/com/samsara/api/resources/webhooks/requests/ListWebhooksRequest.java
@@ -22,7 +22,7 @@ import java.util.Optional;
 public final class ListWebhooksRequest {
     private final Optional<String> ids;
 
-    private final Optional<Integer> limit;
+    private final Optional<Long> limit;
 
     private final Optional<String> after;
 
@@ -30,7 +30,7 @@ public final class ListWebhooksRequest {
 
     private ListWebhooksRequest(
             Optional<String> ids,
-            Optional<Integer> limit,
+            Optional<Long> limit,
             Optional<String> after,
             Map<String, Object> additionalProperties) {
         this.ids = ids;
@@ -51,7 +51,7 @@ public final class ListWebhooksRequest {
      * @return The limit for how many objects will be in the response. Default and max for this value is 512 objects.
      */
     @JsonProperty("limit")
-    public Optional<Integer> getLimit() {
+    public Optional<Long> getLimit() {
         return limit;
     }
 
@@ -96,7 +96,7 @@ public final class ListWebhooksRequest {
     public static final class Builder {
         private Optional<String> ids = Optional.empty();
 
-        private Optional<Integer> limit = Optional.empty();
+        private Optional<Long> limit = Optional.empty();
 
         private Optional<String> after = Optional.empty();
 
@@ -130,12 +130,12 @@ public final class ListWebhooksRequest {
          * <p>The limit for how many objects will be in the response. Default and max for this value is 512 objects.</p>
          */
         @JsonSetter(value = "limit", nulls = Nulls.SKIP)
-        public Builder limit(Optional<Integer> limit) {
+        public Builder limit(Optional<Long> limit) {
             this.limit = limit;
             return this;
         }
 
-        public Builder limit(Integer limit) {
+        public Builder limit(Long limit) {
             this.limit = Optional.ofNullable(limit);
             return this;
         }
